+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for other entrypoints other than `/bin/sh`, can specify an entrypoint for the shell command method.
+- Add logging library `glog` to allow logging at levels of severity and verbosity.
+- Can specify verbosity level of logs via the `logVerbosity` configuration option.
 ### Changed
 - Can scale ReplicaSets, ReplicationControllers and StatefulSets alongside Deployments.
 - ResourceMetrics fields have `resourceName` and `resource` rather than `deploymentName` and `deployment`. In JSON this means that only the resource name will be exposed via field `resource`.